♻️ context 리팩토링 적용
